requirements.txt
selenium
tqdm
pandas
numpy
argparse
urllib3
BeautifulSoup4
wordcloud
matplotlib
Pillow